Cuda updates

+++ libnd4j/include/ops/declarable/helpers/cuda/convolutions_conv2dBP.cu
@@ -35,7 +35,7 @@ namespace ops {
 template <typename X, typename Y>
 static void conv2dBP_(sd::graph::Context& block, const NDArray* input, const NDArray* weights, const NDArray* bias,
                       const NDArray* gradO, NDArray* gradI, NDArray* gradW, NDArray* gradB, const int kH, const int kW,
-                      const int sH, const int sW, int pH, int pW, const int dH, const int dW, const int paddingMode,
+                      const int sH, const int sW, LongType pH, LongType pW, const int dH, const int dW, const int paddingMode,
                       const int isNCHW, const int wFormat) {
   // input   [bS, iH, iW, iC] (NHWC) or [bS, iC, iH, iW] (NCHW)
   // weights [kH, kW, iC, oC], [oC, iC, kH, kW], [oC, kH, kW, iC]
