+++ libnd4j/include/ops/declarable/helpers/cuda/convolutions_depthwiseConv2d.cu
@@ -35,7 +35,7 @@ namespace ops {
 template <typename X, typename Y>
 static void depthwiseConv2d_(sd::graph::Context& block, const NDArray* input, const NDArray* weights,
                              const NDArray* bias, NDArray* output, const int kH, const int kW, const int sH,
-                             const int sW, int pH, int pW, const int dH, const int dW, const int paddingMode,
+                             const int sW, LongType pH, LongType pW, const int dH, const int dW, const int paddingMode,
                              const int isNCHW, const int wFormat) {
   // input     [bS, iH, iW, iC] (NHWC) or [bS, iC, iH, iW] (NCHW)
   // weights   [kH, kW, iC, mC], [mC, iC, kH, kW], [mC, kH, kW, iC]
